Revert "Added tests for api"
