+++ .circleci/config.yml
@@ -17,23 +17,14 @@ jobs:
           docker build -t jeverd/liteboard-ngnix .
           echo $DOCKER_PASS | docker login -u $DOCKER_USER --password-stdin  # (4)
           docker push jeverd/liteboard-ngnix 
-  tests:
-    docker:
-      - image: circleci/node:10.15.0
-      - image: redis
-    steps:
-      - checkout
-      - setup_remote_docker
-      - run:
-          name: 'Running tests'
-          command: |
-            echo 'running npm install'
-            npm install
-            cd config
-            mv example.env .env
-            echo 'now running tests'
-            cd ..
-            npm test
+          
+      # - run:
+      #     name: 'Deploy to AWS EC2 INSTANACE'
+      #     command: |    
+      #       echo 'Deploying to EC2'
+      #       echo 'ec2-35-183-20-250.ca-central-1.compute.amazonaws.com' >> ~/.ssh/known_hosts
+      #       ssh -o StrictHostKeyChecking=no ec2-user@ec2-35-183-20-250.ca-central-1.compute.amazonaws.com "./deploy.sh"
+
 
 workflows:
   version: 2
@@ -45,5 +36,3 @@ workflows:
               only: 
                 - dev
                 - dockerUPdate
-      - tests
-
