+++ config/config.js
@@ -14,5 +14,4 @@ module.exports = {
   email: process.env.EMAIL,
   emailPassword: process.env.EMAIL_PASSWORD,
   emailService: process.env.EMAIL_SERVICE,
-  loggerFlag: (process.env.LOGGER === 'true'),
 };
