+++ config/example.env
@@ -1,7 +1,7 @@
 REDIS_HOST=127.0.0.1
 REDIS_PORT=6379
 EXPRESS_PORT=8080
-NODE_ENV=DEVELOPMENT
+NODE_ENV=PRODUCTION
 PT_HOST=logs6.papertrailapp.com
 PT_PORT=29341
 WEB_SERVER_PORT_HTTP=80 
@@ -10,4 +10,3 @@ REDIS_URL=redis://redis_db:6379
 EMAIL=ece150sucks@gmail.com
 EMAIL_PASSWORD=jawad123
 EMAIL_SERVICE=Gmail
-LOGGER=true # usually only change this to false when you want to run tests
