+++ package.json
@@ -5,7 +5,7 @@
   "main": "servers.js",
   "scripts": {
     "start": "node index.js",
-    "test": "set LOGGER=false && mocha --exit"
+    "test": "mocha"
   },
   "repository": {
     "type": "git",
@@ -33,7 +33,6 @@
     "winston-papertrail": "git+https://github.com/kenperkins/winston-papertrail.git#v2"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.1.0",
     "eslint-config-airbnb-base": "^14.1.0",
