+++ server/servers.js
@@ -7,7 +7,7 @@ const socketio = require('socket.io');
 const bodyParser = require('body-parser');
 
 const {
-  redisHost, redisPort, expressPort, environment, redisUrl, loggerFlag,
+  redisHost, redisPort, expressPort, environment, redisUrl,
 } = require('../config/config');
 const { logger } = require('./services/logger/logger');
 const { logMiddleWare } = require('./services/logger/loggingMiddleware');
@@ -17,7 +17,6 @@ const app = express();
 const expressServer = app.listen(expressPort);
 const io = socketio(expressServer);
 const peerServer = ExpressPeerServer(expressServer);
-
 app.use('/peerjs', peerServer);
 app.use(express.static('public/js'));
 app.use(express.static('public/css'));
@@ -25,7 +24,7 @@ app.use(express.static('public/images'));
 app.use(express.json({ limit: '50mb' }));
 app.use(bodyParser.json());
 app.use(helmet());
-if (loggerFlag) app.use(logMiddleWare);
+app.use(logMiddleWare);
 
 
 let client = null;
